Feature/aidt 8 red target definition

+++ yawning_titan/envs/generic/core/action_loops.py
@@ -3,10 +3,14 @@ The ``ActionLoop`` class helps reduce boilerplate code when evaluating an agent 
 
 Serves a similar function to library helpers such as Stable Baselines 3 ``evaluate_policy()".
 """
+
+from datetime import datetime
 import os
 
 import imageio
 import matplotlib.pyplot as plt
+import pandas as pd
+
 
 from yawning_titan import IMAGES_DIR
 
@@ -29,7 +33,7 @@ class ActionLoop:
         self.filename = filename
         self.episode_count = episode_count
 
-    def gif_action_loop(self, render: bool = True):
+    def gif_action_loop(self,render_network=True,prompt_to_close=False,save_gif=False,deterministic=True,*args,**kwargs):
         """
         Run the agent in evaluation and create a gif from episodes.
 
@@ -41,7 +45,9 @@ class ActionLoop:
             # if the path does not exist, create it
             os.makedirs(IMAGES_DIR)
 
+        complete_results = []
         for i in range(self.episode_count):
+            results = pd.DataFrame(columns = ["action","rewards","info"]) # temporary log to satisfy repeatability tests until logging can be full implemented
             obs = self.env.reset()
             done = False
             frame_names = []
@@ -49,59 +55,82 @@ class ActionLoop:
 
             while done is False:
                 # gets the agents prediction for the best next action to take
-                action, _states = self.agent.predict(obs, deterministic=True)
+                action, _states = self.agent.predict(obs, deterministic=deterministic)
+                
                 # TODO: setup logging properly here
                 # logging.info(f'Blue Agent Action: {action}')
                 # step the env
                 obs, rewards, done, info = self.env.step(action)
 
+                results.loc[len(results.index)] = [action,rewards,info]
+
                 # TODO: setup logging properly here
                 # logging.info(f'Observations: {obs.flatten()} Rewards:{rewards} Done:{done}')
                 # self.env.render(episode=i+1)
-                if render:
-                    self.env.render()
 
-                current_name = os.path.join(
-                    IMAGES_DIR, f"image_{current_image}.png"
+                if save_gif:
+                    current_name = os.path.join(
+                        IMAGES_DIR, f"image_{current_image}.png"
+                    )
+                    current_image += 1
+                    frame_names.append(current_name)
+                    # save the current image
+                    plt.savefig(current_name)
+
+                    current_image += 1
+                    frame_names.append(current_name)
+                    # save the current image
+                    plt.savefig(current_name)
+
+
+                if render_network:
+                    self.env.render(*args,**kwargs)       
+
+            if save_gif:
+                string_time = datetime.now().strftime("%d%m%Y_%H%M%S")
+                gif_path = os.path.join(
+                    IMAGES_DIR, f"{self.filename}_{string_time}_{self.episode_count}.gif"
                 )
-                current_image += 1
-                frame_names.append(current_name)
-                # save the current image
-                plt.savefig(current_name)
+                with imageio.get_writer(gif_path, mode="I") as writer:
+                    # create a gif from the images
+                    for filename in frame_names:
+                        image = imageio.imread(filename)
+                        writer.append_data(image)
 
-            gif_path = os.path.join(
-                IMAGES_DIR, f"{self.filename}_{self.episode_count}.gif"
-            )
-            with imageio.get_writer(gif_path, mode="I") as writer:
-                # create a gif from the images
-                for filename in frame_names:
-                    image = imageio.imread(filename)
-                    writer.append_data(image)
+                    for filename in set(frame_names):
+                        os.remove(filename)
 
-            for filename in set(frame_names):
-                os.remove(filename)
+            complete_results.append(results)
 
-        self.env.close()
+        if not prompt_to_close:
+            self.env.close()
+        return complete_results
 
-    def standard_action_loop(self):
-        """Indefintely act within the environment using a trained agent."""
+    def standard_action_loop(self,deterministic=False):
+        """Indefinitely act within the environment using a trained agent."""
+        complete_results = []
         for i in range(self.episode_count):
+            results = pd.DataFrame(columns = ["action","rewards","info"]) # temporary log to satisfy repeatability tests until logging can be full implemented
             obs = self.env.reset()
             done = False
             while not done:
-                action, _states = self.agent.predict(obs)
+                action, _states = self.agent.predict(obs,deterministic=deterministic)
                 # TODO: setup logging properly here
                 # logging.info(f'Blue Agent Action: {action}')
                 obs, rewards, done, info = self.env.step(action)
+                results.loc[len(results.index)] = [action,rewards,info]
+            complete_results.append(results)
+        return complete_results
+            
 
-    def random_action_loop(self):
-        """Indefintely act within the environment taking random actions."""
+    def random_action_loop(self,deterministic=False):
+        """Indefinitely act within the environment taking random actions."""
         for i in range(self.episode_count):
             obs = self.env.reset()
             done = False
             reward = 0
             while not done:
-                action = self.agent.predict(obs, reward, done)
+                action = self.agent.predict(obs, reward, done,deterministic=deterministic)
                 ob, reward, done, ep_history = self.env.step(action)
                 if done:
                     break
